+++ components/widgets/psi/index.js
@@ -1,7 +1,7 @@
 import { Component } from 'react'
 import { URL } from 'universal-url'
 import fetch from 'isomorphic-unfetch'
-import Progress from '../../progress'
+import CircleProgress from '../../circle-progress'
 import Widget from '../../widget'
 
 export default class PageSpeedInsights extends Component {
@@ -34,7 +34,7 @@ export default class PageSpeedInsights extends Component {
     const { score } = this.state
     return (
       <Widget title='PageSpeed Score'>
-        <Progress value={score} />
+        <CircleProgress value={score} />
       </Widget>
     )
   }
